Created new resources files and rearranged resources to the corresponding files. Finished first version of the Senior Player Grid. Added new BusinessObjects entities, enums, constants, mappings, managers, related factories, strategies and extension methods. Deleted SortableBindingList since it was useless. Changed a few Data Types in the SeniorPlayerWithSkillDelta entity and view. Created new controls. Updated 001_SeniorPlayerWithSkillDelta.sql and 002_Grids.sql scripts.

+++ Hyperar.HattrickUltimate.BusinessObjects/App/SeniorPlayerWithSkillDelta.cs
@@ -23,12 +23,12 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.App
         /// <summary>
         /// Gets or sets the Aggressiveness.
         /// </summary>
-        public byte Aggressiveness { get; set; }
+        public int Aggressiveness { get; set; }
 
         /// <summary>
         /// Gets or sets the Agreeability.
         /// </summary>
-        public byte Agreeability { get; set; }
+        public int Agreeability { get; set; }
 
         /// <summary>
         /// Gets or sets the Booking Status.
@@ -131,7 +131,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.App
         /// <summary>
         /// Gets or sets the Honesty.
         /// </summary>
-        public byte Honesty { get; set; }
+        public int Honesty { get; set; }
 
         /// <summary>
         /// Gets or sets the Injury Status.
@@ -161,7 +161,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.App
         /// <summary>
         /// Gets or sets the Leadership.
         /// </summary>
-        public byte Leadership { get; set; }
+        public int Leadership { get; set; }
 
         /// <summary>
         /// Gets or sets the Loyalty.
@@ -241,7 +241,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.App
         /// <summary>
         /// Gets or sets the Specialty.
         /// </summary>
-        public byte Specialty { get; set; }
+        public int Specialty { get; set; }
 
         /// <summary>
         /// Gets or sets the Stamina.
